+++ src/reporter/TextReporter.php
@@ -33,9 +33,6 @@ class TextReporter implements ReporterInterface
     const DEFAULT_LOW_BOUND = 35.0;
     const DEFAULT_HIGH_BOUND = 70.0;
 
-    const PAD_CHARACTER = '.';
-    const PAD_CHARACTER_LENGTH = 70;
-
     /**
      * @var \cloak\writer\ConsoleWriter
      */
@@ -83,13 +80,13 @@ class TextReporter implements ReporterInterface
      */
     public function reportResult(Result $result)
     {
-        $this->writeTotalCoverage($result);
-
         $files = $result->getFiles()->getIterator();
 
         foreach ($files as $file) {
             $this->reportFile($file);
         }
+
+        $this->writeTotalCoverage($result);
     }
 
     /**
@@ -99,15 +96,16 @@ class TextReporter implements ReporterInterface
     {
         $currentDirectory = getcwd();
 
-        $filePathReport = $file->getRelativePath($currentDirectory) . ' ';
-        $filePathReport = str_pad($filePathReport, static::PAD_CHARACTER_LENGTH, static::PAD_CHARACTER);
+        $filePathReport = $file->getRelativePath($currentDirectory);
 
-        $this->console->writeText($filePathReport);
-        $this->writeFileCoverage($file);
+        $this->writeCoverage($file);
+        $this->console->writeText(' ');
         $this->console->writeText(sprintf("(%2d/%2d)",
             $file->getExecutedLineCount(),
             $file->getExecutableLineCount()
         ));
+        $this->console->writeText(' ');
+        $this->console->writeText($filePathReport);
 
         $this->console->writeText(PHP_EOL);
     }
@@ -117,21 +115,15 @@ class TextReporter implements ReporterInterface
      */
     protected function writeTotalCoverage(Result $result)
     {
-        $this->console->writeText('Total code coverage:');
+        $this->console->writeText(PHP_EOL);
+        $this->console->writeText('Code Coverage:');
         $this->writeCoverage($result);
-        $this->console->writeText(PHP_EOL . PHP_EOL);
+        $this->console->writeText(PHP_EOL);
     }
 
     /**
      * @param \cloak\result\File $file
      */
-    protected function writeFileCoverage(File $file)
-    {
-        $this->console->writeText(' ');
-        $this->writeCoverage($file);
-        $this->console->writeText(' ');
-    }
-
     protected function writeCoverage(CoverageResultInterface $result)
     {
         $text = sprintf('%6.2f%%', $result->getCodeCoverage()->value());