+++ src/writer/ConsoleWriter.php
@@ -18,20 +18,20 @@ use Zend\Console\Adapter\AdapterInterface;
  * Class ConsoleWriter
  * @package cloak\writer
  */
-class ConsoleWriter
+class ConsoleWriter implements WriterInterface
 {
 
     /**
      * @var \Zend\Console\Adapter\AdapterInterface
      */
-    private $cosnole;
+    private $console;
 
     /**
      * @param AdapterInterface $adapterInterface
      */
     public function __construct(AdapterInterface $adapterInterface = null)
     {
-        $this->cosnole = $adapterInterface ?: Console::getInstance();
+        $this->console = $adapterInterface ?: Console::getInstance();
     }
 
     /**
@@ -44,7 +44,7 @@ class ConsoleWriter
      */
     public function writeText($text, $color = null, $bgColor = null)
     {
-        $this->cosnole->writeText($text, $color, $bgColor);
+        $this->console->writeText($text, $color, $bgColor);
     }
 
     /**
@@ -58,12 +58,15 @@ class ConsoleWriter
      */
     public function writeLine($text = "", $color = null, $bgColor = null)
     {
-        $this->cosnole->writeLine($text, $color, $bgColor);
+        $this->console->writeLine($text, $color, $bgColor);
     }
 
+    /**
+     * Write a blank line
+     */
     public function writeEOL()
     {
-        $this->cosnole->writeLine('');
+        $this->console->writeLine('');
     }
 
 }