+++ src/writer/FileWriter.php
@@ -17,7 +17,7 @@ use SplFileObject;
  * Class FileWriter
  * @package cloak\writer
  */
-class FileWriter
+class FileWriter implements WriterInterface
 {
 
     /**
@@ -33,11 +33,11 @@ class FileWriter
         $directoryPath = dirname($filePath);
 
         if (file_exists($directoryPath) === false) {
-            throw new DirectoryNotFoundException();
+            throw new DirectoryNotFoundException("$directoryPath directory not found");
         }
 
         if (is_writable($directoryPath) === false) {
-            throw new DirectoryNotWritableException();
+            throw new DirectoryNotWritableException("Can not write to the directory $directoryPath");
         }
 
         $this->file = new SplFileObject($filePath, 'w');
@@ -57,6 +57,14 @@ class FileWriter
     public function writeLine($text)
     {
         $this->file->fwrite($text . PHP_EOL);
+    }
+
+    /**
+     * Write a blank line
+     */
+    public function writeEOL()
+    {
+        $this->file->fwrite(PHP_EOL);
     }
 
     /**